pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tokioai"
version = "2.0.0"
description = "TokioAI — Autonomous AI Agent for Security and Infrastructure"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "TokioAI"}]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.43.0",
    "openai>=1.60.0",
    "google-generativeai>=0.8.0",
    "psycopg2-binary>=2.9.9",
    "aiohttp>=3.10.0",
    "httpx>=0.27.0",
    "docker>=7.1.0",
    "pyyaml>=6.0",
    "rich>=13.9.0",
]

[project.scripts]
tokio = "tokio_agent.cli:main"

[tool.setuptools.packages.find]
include = ["tokio_agent*"]
